Don't swap two guesses
If current guess was guessed elsewhere, and current highlighted square
already contains a guess, do not input the guess of the highlighted
square as a guess for where the current guess was previously.
Instead, treat previously guessed square for current guess as if it had
no guess.

Update planning docs